refactor: split listeners
