+++ src/main.rs
@@ -3,37 +3,31 @@
 
 mod command;
 mod model;
+mod listener;
 
 use std::collections::HashMap;
-use std::fs::{self, File};
-use std::io::{self, BufReader, Write};
-use std::path::{Path, PathBuf};
-use std::sync::{Arc, Mutex};
+use std::fs::File;
+use std::io::{BufReader, Write};
+use std::sync::Mutex;
 
 use once_cell::sync::OnceCell;
-use reqwest::header::CONTENT_TYPE;
 use serenity::{
-    async_trait,
-    client::{Client, Context, EventHandler},
+    client::Client,
     framework::StandardFramework,
     model::{
         channel::Message,
-        gateway::Ready,
         prelude::{ChannelId, GuildId},
     },
     Result as SerenityResult,
 };
-use songbird::{
-    ffmpeg, tracks::create_player, Event, EventContext,
-    EventHandler as VoiceEventHandler, SerenityInit, Songbird, TrackEvent,
-};
-use uuid::Uuid;
+use songbird::SerenityInit;
 use crate::{
     command::GENERAL_GROUP,
     model::{
         State, Config, UserSettings
     }
 };
+use crate::listener::serenity::Handler;
 
 #[macro_use]
 extern crate lazy_static;
@@ -47,168 +41,6 @@ lazy_static! {
 }
 
 static CONFIG: OnceCell<Config> = OnceCell::new();
-
-struct Handler;
-
-#[async_trait]
-impl EventHandler for Handler {
-    async fn ready(&self, _: Context, ready: Ready) {
-        println!("{} is connected!", ready.user.name);
-    }
-
-    async fn message(&self, ctx: Context, msg: Message) {
-        if msg.author.bot {
-            return;
-        }
-
-        if msg.content == "ping" {
-            check_msg(msg.channel_id.say(&ctx.http, "[discord-tts] pong").await);
-            return;
-        }
-
-        match msg.content.get(..1) {
-            Some("~") => return,
-            Some(";") => match msg.content.chars().nth(1) {
-                Some(';') => {}
-                _ => return,
-            },
-            _ => {}
-        };
-
-        let guild_id = match msg.guild_id {
-            Some(guild_id) => guild_id,
-            None => return,
-        };
-
-        let manager = songbird::get(&ctx)
-            .await
-            .expect("Songbird Voice client placed in at initialisation.")
-            .clone();
-
-        let handler = match manager.get(guild_id) {
-            Some(handler) => handler,
-            None => return,
-        };
-
-        {
-            let m = CURRENT_TEXT_CHANNEL.lock().unwrap();
-            match m.get(&guild_id) {
-                Some(channel_id) => {
-                    if channel_id != &msg.channel_id {
-                        return;
-                    }
-                }
-                None => return,
-            }
-        }
-
-        let speaker = {
-            let s = STATE.lock().unwrap();
-            match s.user_settings.get(&msg.author.id) {
-                Some(setting) => match setting.speaker {
-                    Some(speaker) => speaker,
-                    _ => 0,
-                },
-                None => 0,
-            }
-            .to_string()
-        };
-
-        let c = CONFIG.get().unwrap();
-
-        let params = [("text", msg.content.as_str()), ("speaker", &speaker)];
-        let client = reqwest::Client::new();
-        let query = client
-            .post(format!("{}/audio_query", c.voicevox_host))
-            .query(&params)
-            .send()
-            .await
-            .expect("Failed to create audio query");
-
-        let query = query.text().await.expect("Failed to get text");
-
-        let params = [("speaker", &speaker)];
-        let audio = client
-            .post(format!("{}/synthesis", c.voicevox_host))
-            .query(&params)
-            .header(CONTENT_TYPE, "application/json")
-            .body(query)
-            .send()
-            .await
-            .expect("Failed to create audio query");
-
-        let uuid = Uuid::new_v4().to_string();
-        let path = Path::new(&c.tmp_path).join(&uuid);
-
-        let mut output = File::create(&path).expect("Failed to create file");
-        let audio = audio.bytes().await.expect("Failed to read resp");
-        let mut response_cursor = std::io::Cursor::new(audio);
-        io::copy(&mut response_cursor, &mut output).expect("Failed to write file");
-
-        let mut handler = handler.lock().await;
-
-        let source = match ffmpeg(&path).await {
-            Ok(source) => source,
-            Err(why) => {
-                println!("Err starting source: {:?}", why);
-                check_msg(msg.reply(ctx, "Error sourcing ffmpeg").await);
-                return;
-            }
-        };
-
-        let (audio, audio_handle) = create_player(source);
-
-        audio_handle
-            .add_event(
-                Event::Track(TrackEvent::End),
-                ReadEndNotifier {
-                    temporary_filename: path,
-                },
-            )
-            .expect("Failed to create queue");
-
-        handler.enqueue(audio);
-    }
-}
-
-struct ReadEndNotifier {
-    temporary_filename: PathBuf,
-}
-
-#[async_trait]
-impl VoiceEventHandler for ReadEndNotifier {
-    async fn act(&self, ctx: &EventContext<'_>) -> Option<Event> {
-        if let EventContext::Track(_) = ctx {
-            fs::remove_file(&self.temporary_filename).expect("Failed to remove temporary file")
-        }
-        None
-    }
-}
-
-struct DriverDisconnectNotifier {
-    songbird_manager: Arc<Songbird>,
-}
-
-#[async_trait]
-impl VoiceEventHandler for DriverDisconnectNotifier {
-    async fn act(&self, ctx: &EventContext<'_>) -> Option<Event> {
-        if let EventContext::DriverDisconnect(ctx) = ctx {
-            let guild_id = ctx.guild_id;
-            let manager = &self.songbird_manager;
-            let has_handler = manager.get(guild_id).is_some();
-
-            println!("Force disconnected");
-
-            if has_handler {
-                manager
-                    .remove(guild_id)
-                    .await
-                    .expect("Failed to remove from manager");
-            }
-        }
-        None
-    }
-}
 
 #[tokio::main]
 async fn main() {
